backend/requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
Flask-Mail==0.9.1
requests==2.28.1
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
beautifulsoup4==4.11.1
itsdangerous==2.1.2
schedule==1.2.1
SQLAlchemy==2.0.25
watchdog==3.0.0
Werkzeug==3.0.1
python-dotenv==0.20.0
unittest2==1.1.0

